+++ packages/axway-cli-oum/src/team/user/add.js
@@ -45,7 +45,8 @@ export default {
 		},
 		'--role <role>': {
 			desc: 'Assign one or more team roles to a user',
-			multiple: true
+			multiple: true,
+			redact: false
 		}
 	},
 	async action({ argv, cli, console }) {